Crate scattering

Reexports

pub use material::Material;

pub use stats::Stats;

pub use fields::Fields;

pub use probability::probability;

Modules

boltzmann
Provides function for creating ensembles of particles with Boltzmann distribution

fields

material

particle
Particle one particle movement in material under electromagnetic fields with phonon scattering

probability
Provides function for calculate probability of electron-phonon scattering

stats

Functions

create_ensemble